Add the option to send a message to the server
Input/output variables for sending and receiving messages have been
created in Client class. For the input there is a variable of
BufferedReader class, for output - PrintWriter' instance.
Added option of closing a connection by send to the server a "bye" message.
Input stream from socket displays an information if the server received
a message, sent by a client.
If there is not connection to the server, it will displays a comunicat
that the server is unavailable.